[#39] move blueprints to common
